Added caching for generated API
