+++ www/data/api.ts
@@ -1,23 +1,5 @@
-import { doc, load } from "../utils/doc.ts";
+import RAW_API from "../../docs/api.json" with { type: "json" };
 
-const modules = ["mod", "namespaces", "rdf", "sparql"];
+export type Api = typeof RAW_API;
 
-async function loadDoc(module: string) {
-  const file = import.meta.resolve(`../../${module}.ts`);
-  return await doc(file, { printImportMapDiagnostics: true, load });
-}
-
-export async function getApi() {
-  const promises = modules.map(async (module) => {
-    const x = await loadDoc(module);
-    return {
-      kind: "module",
-      path: `/${module}.ts`,
-      items: x,
-    };
-  });
-
-  return await Promise.all(promises);
-}
-
-export const api = await getApi();
+export const api: Api = RAW_API;
