+++ www/routes/_app.tsx
@@ -1,0 +1,16 @@
+import { PageProps } from "$fresh/server.ts";
+
+export default function App({ Component }: PageProps) {
+  return (
+    <html>
+      <head>
+        <meta charset="utf-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>www</title>
+      </head>
+      <body>
+        <Component />
+      </body>
+    </html>
+  );
+}
